+++ backend/internal/service/admin_service.go
@@ -2740,4 +2740,3 @@ func (s *adminServiceImpl) ForceAntigravityPrivacy(ctx context.Context, account 
 	applyAntigravityPrivacyMode(account, mode)
 	return mode
 }
-
