+++ backend/internal/service/antigravity_oauth_service.go
@@ -322,7 +322,7 @@ func (s *AntigravityOAuthService) RefreshAccountToken(ctx context.Context, accou
 // loadCodeAssistResult 封装 loadProjectIDWithRetry 的返回结果，
 // 同时携带从 LoadCodeAssist 响应中提取的 plan_type 信息。
 type loadCodeAssistResult struct {
-	ProjectID string
+	ProjectID    string
 	Subscription *AntigravitySubscriptionResult
 }
