+++ backend/internal/service/token_refresh_service.go
@@ -526,4 +526,3 @@ func (s *TokenRefreshService) ensureAntigravityPrivacy(ctx context.Context, acco
 		)
 	}
 }
-
